fixed the naming schema and .cjs to .js issue
