+++ backend/scripts/parse.js
@@ -1,4 +1,4 @@
-const fs = require('fs');
+import fs from 'fs';
 
 function parseTestCases(block) {
     if (!block) return [];
